refactor: 세션 정보를 request 객체에 넘겨주도록 변경
- 반복되는 세션 조회 코드 제거를 위해 변경
