+++ src/main/java/codesquad/command/CommandManager.java
@@ -133,11 +133,20 @@ public class CommandManager {
 		checkHandling(path, method);
 
 
+		var httpClientRequest = new HttpClientRequest(httpRequest);
+		var httpClientResponse = new HttpClientResponse();
+
 		// 실패하면 /index.html로 리다이렉트
 		if (!callInterceptor(path, httpRequest)) {
-			var httpClientResponse = new HttpClientResponse();
-			httpClientResponse.setHeader("Location","/login/index.html");
-			return new DomainResponse(HttpStatus.FOUND,  httpClientResponse, false, method.getReturnType(), null);
+			httpClientResponse.setHeader("Location", "/login/index.html");
+			return new DomainResponse(HttpStatus.FOUND, httpClientResponse, false, method.getReturnType(), null);
+		} else {
+			var cookieInfo = httpRequest.cookie();
+			Cookie cookie = cookieInfo.get("sessionKey");
+			if (Objects.nonNull(cookie)) {
+				SessionUserInfo userInfo = Session.getInstance().getSession(cookie.value());
+				httpClientRequest.setUserInfo(userInfo);
+			}
 		}
 		log.info("[Execute Method] : , {}",method);
 
@@ -145,8 +154,6 @@ public class CommandManager {
 			var className = method.getDeclaringClass().getName();
 			var instance = findInstance(className);
 
-			var httpClientRequest = new HttpClientRequest(httpRequest);
-			var httpClientResponse = new HttpClientResponse();
 			var userInputData = parsingQueryParameterResources(resources);
 			var parameters = makeParameterArgs(method, userInputData, httpClientRequest, httpClientResponse);
 			log.info("[User Parameters] : {}", Arrays.toString(parameters));
