+++ src/main/java/codesquad/command/domainResponse/HttpClientRequest.java
@@ -2,6 +2,7 @@ package codesquad.command.domainResponse;
 
 import codesquad.http.request.format.HttpRequest;
 import codesquad.session.Cookie;
+import codesquad.session.SessionUserInfo;
 
 import java.util.Map;
 
@@ -9,6 +10,7 @@ public class HttpClientRequest {
     Map<String,String> headers;
     Map<String, Cookie> cookies;
     String uri;
+    SessionUserInfo userInfo;
 
     public HttpClientRequest(HttpRequest httpRequest) {
         this.headers = httpRequest.headers();
@@ -18,5 +20,12 @@ public class HttpClientRequest {
 
     public Cookie getCookie(String cookieName) {
         return cookies.get(cookieName);
+    }
+
+    public SessionUserInfo getUserInfo() {
+        return this.userInfo;
+    }
+    public void setUserInfo(SessionUserInfo userInfo) {
+        this.userInfo = userInfo;
     }
 }
