+++ src/test/java/codesquad/command/domain/member/MemberDomainTest.java
